+++ src/test/java/no/nav/pto/veilarbportefolje/arenapakafka/TiltakPostgresTest.java
@@ -11,7 +11,7 @@ import no.nav.pto.veilarbportefolje.database.PostgresTable;
 import no.nav.pto.veilarbportefolje.domene.EnhetTiltak;
 import no.nav.pto.veilarbportefolje.domene.value.PersonId;
 import no.nav.pto.veilarbportefolje.oppfolgingsbruker.OppfolgingsbrukerEntity;
-import no.nav.pto.veilarbportefolje.oppfolgingsbruker.OppfolgingsbrukerRepository;
+import no.nav.pto.veilarbportefolje.oppfolgingsbruker.OppfolgingsbrukerRepositoryV2;
 import no.nav.pto.veilarbportefolje.postgres.AktivitetOpensearchService;
 import no.nav.pto.veilarbportefolje.postgres.utils.AvtaltAktivitetEntity;
 import no.nav.pto.veilarbportefolje.postgres.AktivitetEntityDto;
@@ -34,7 +34,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest(classes = ApplicationConfigTest.class)
 public class TiltakPostgresTest {
     private final JdbcTemplate jdbcTemplatePostgres;
-    private final OppfolgingsbrukerRepository oppfolgingsbrukerRepository;
+    private final OppfolgingsbrukerRepositoryV2 oppfolgingsbrukerRepositoryV2;
     private final TiltakRepositoryV2 tiltakRepositoryV2;
     private final AktivitetOpensearchService aktivitetOpensearchService;
 
@@ -43,9 +43,9 @@ public class TiltakPostgresTest {
     private final PersonId personId = PersonId.of("123");
 
     @Autowired
-    public TiltakPostgresTest(@Qualifier("PostgresJdbc") JdbcTemplate jdbcTemplatePostgres, TiltakRepositoryV2 tiltakRepositoryV2, AktivitetOpensearchService aktivitetOpensearchService, OppfolgingsbrukerRepository oppfolgingsbrukerRepository) {
+    public TiltakPostgresTest(@Qualifier("PostgresJdbc") JdbcTemplate jdbcTemplatePostgres, TiltakRepositoryV2 tiltakRepositoryV2, AktivitetOpensearchService aktivitetOpensearchService, OppfolgingsbrukerRepositoryV2 oppfolgingsbrukerRepositoryV2) {
         this.jdbcTemplatePostgres = jdbcTemplatePostgres;
-        this.oppfolgingsbrukerRepository = oppfolgingsbrukerRepository;
+        this.oppfolgingsbrukerRepositoryV2 = oppfolgingsbrukerRepositoryV2;
         this.aktivitetOpensearchService = aktivitetOpensearchService;
         this.tiltakRepositoryV2 = tiltakRepositoryV2;
     }
@@ -161,7 +161,7 @@ public class TiltakPostgresTest {
     @Test
     public void skal_lagre_tiltak_pa_enhet() {
         String navKontor = "0007";
-        oppfolgingsbrukerRepository.leggTilEllerEndreOppfolgingsbruker(
+        oppfolgingsbrukerRepositoryV2.leggTilEllerEndreOppfolgingsbruker(
                 new OppfolgingsbrukerEntity(aktorId.get(), null, null, null, "" +
                         "Tester", "Testerson", navKontor, null, null, null, null,
                         "1234", true, true, false, null, ZonedDateTime.now()));
